+++ saturn/ovl.h
@@ -12,7 +12,7 @@ extern Uint32 *shared;
 //static int nBurnSoundLen;
 //static Sint8 *nSoundBuffer = (Sint8 *)0x25a20000;
 extern unsigned short nSoundBufferPos;
-extern unsigned short nBurnSprites;
+extern unsigned int nBurnSprites;
 
 #define	SZ_PERIPHERAL	20
 typedef	UINT8	SysPeripheral[SZ_PERIPHERAL+2];
